RIL:add the ril wrapper dmitry-ril
Samsung's ancient RIL blob would crash instead of reporting unknown when
receiving some unhandled requests. Add a wrapper won't touch the souce
and would be easier to maintain in the future.
Should there be an updated blob, we can drop this wrapper.

Wrapper library is not a new thing but I found it in Dirmtry Grinbery's
site. Keeping the name to reflect his credit seems appropriate. Minor
modifications were made.